Move board-related components under [boardId] folder
