+++ app/(user)/server/[serverId]/boards/[boardId]/_components/Frames/FrameStyle1.tsx
