+++ app/(user)/server/[serverId]/boards/[boardId]/_components/Frames/FrameStyle2.tsx
