+++ app/(user)/server/[serverId]/boards/[boardId]/_components/Frames/FrameStyle3.tsx
