+++ app/(user)/server/[serverId]/boards/[boardId]/_components/GamePiece.tsx
@@ -1,10 +1,10 @@
 "use client";
 import ColumnWrapper from "@/app/_components/wrappers/ColumnWrapper";
 import { useEffect, useState } from "react";
+import getBlobSASUrl from "@/actions/getBlobSASUrl";
 import FrameStyle1 from "./Frames/FrameStyle1";
 import FrameStyle2 from "./Frames/FrameStyle2";
 import FrameStyle3 from "./Frames/FrameStyle3";
-import getBlobSASUrl from "@/actions/getBlobSASUrl";
 
 export default function GamePiece({
   character,
