+++ app/(user)/server/[serverId]/boards/[boardId]/_components/GamePieceBoardWrapper.tsx
