+++ app/(user)/server/[serverId]/boards/[boardId]/_components/ZoomInput.tsx
