+++ app/(user)/server/[serverId]/boards/[boardId]/_components/game piece management/CharacterCarousel.tsx
